feat: выносит фильтр офферов в useFilterOffers
